Chatbot_mémoire_courte/requirements.txt
openai
python-dotenv


